requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
requests==2.32.3
APScheduler==3.10.4
pytz==2025.1
notion-client==2.2.1
google-api-python-client==2.146.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
fastapi==0.115.6
uvicorn==0.30.6
jinja2==3.1.4
pydantic==2.9.2
